fix(playground): render Spec walk as an on-page chip like v52 (#1039)
👤 For humans
Third time's the charm — and the POC already told us how. The v52 version
chip is a plain fixed on-page button in spike-app.vue with a comment saying
exactly this: Nuxt DevTools / the pill "can't show on the mobile staging
preview deploys verify on", so it's rendered on-page. My first two tries
missed that: v1 sat behind app.vue's showChangelogTool gate (false on a
review build), v2 registered into the devtools pill (whose menu only shows
built-in Console/Annotate — custom registerTool doesn't surface). This
renders the "Spec walk" launcher the SAME way as v52: a fixed button right
next to it, ungated. Guaranteed to show because v52 does.

🤖 For agents
- NEW app/components/SpecWalkButton.vue — self-contained fixed launcher
  (bottom-8 left-1, above the v52 chip) + the driver.js walk (spotlight
  [data-handoff] hooks, verdict capture → localStorage, export step).
- spike-app.vue — mounts <SpecWalkButton/> beside the version chip.
- Removed app/plugins/spec-walk.client.ts (registerTool didn't surface).
- Typecheck clean. This is the pattern the POC itself uses for staging.
